배열 스택 top을 의미에 맞게 변경

+++ 01_stack/array_stack.h
@@ -9,7 +9,7 @@ typedef int as_data;
 
 typedef struct {
     as_data *array;
-    size_t top;
+    int top;
     size_t capacity;
 } array_stack;
